Extract IThrowOutQuestionDialogState to its own file
