+++ src/state/DialogState.ts
@@ -12,6 +12,7 @@ import { ModalVisibilityStatus } from "./ModalVisibilityStatus";
 import { RenameTeamDialogState } from "./RenameTeamDialogState";
 import { ImportFromQBJDialogState } from "./ImportFromQBJDialogState";
 import { AddPlayerDialogState } from "./AddPlayerDialogState";
+import { IThrowOutQuestionDialogState } from "./IThrowOutQuestionDialogState";
 import {
     IOKCancelMessageDialogOptions,
     IOKMessageDialogOptions,
@@ -251,15 +252,4 @@ export class DialogState {
     public showNewGameDialog(): void {
         this.visibleDialog = ModalVisibilityStatus.NewGame;
     }
-}
-
-export interface IThrowOutQuestionDialogState {
-    title: string;
-    message: string;
-    // The pre-computed replacement question number (1-based) shown in the SpinButton. Undefined if the packet
-    // has no available replacement (user must upload more questions first).
-    defaultReplacementNumber: number | undefined;
-    // Upper bound for the SpinButton (total questions of this type in the packet).
-    maxQuestionNumber: number;
-    onConfirm: (replacementIndex: number | undefined) => void;
 }
